DC-2129: Moving DomainProfileObjectStore from -api to -impl.  Avoids a potential cyclical dependency, and per a meeting today is not really considered a part of the API.
