+++ dcs-packaging-tool/dcs-packaging-tool-impl/src/main/java/org/dataconservancy/packaging/tool/impl/DomainProfileObjectStore.java
@@ -1,4 +1,4 @@
-package org.dataconservancy.packaging.tool.api;
+package org.dataconservancy.packaging.tool.impl;
 
 import java.net.URI;
 import java.util.List;
